requirements.txt
alembic==1.14.0
bcrypt==4.2.1
bleach==6.2.0
blinker==1.9.0
click==8.1.7
colorama==0.4.6
cssutils==2.11.1
diff-match-patch==20181111
dnspython==2.7.0
email_validator==2.2.0
Flask==3.1.0
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
lxml==5.3.0
Mako==1.3.8
MarkupSafe==3.0.2
more-itertools==10.5.0
mysql-connector-python==9.1.0
passlib==1.7.4
python-dotenv==1.0.1
python-slugify==8.0.4
quill-delta==1.0.3
SQLAlchemy==2.0.36
text-unidecode==1.3
tinycss2==1.4.0
typing_extensions==4.12.2
webencodings==0.5.1
Werkzeug==3.1.3
WTForms==3.2.1
